Add components for 1.21.5, add /loot copycomponents command

+++ Component/v1_21_3/src/main/java/dev/rosewood/roseloot/loot/item/component/v1_21_3/ToolComponent.java
@@ -59,8 +59,8 @@ class ToolComponent implements LootItemComponent {
         Tool.Builder builder = Tool.tool();
 
         if (this.defaultMiningSpeed != null) {
-            double speed = this.defaultMiningSpeed.getDouble(context);
-            builder.defaultMiningSpeed((float) speed);
+            float speed = this.defaultMiningSpeed.getFloat(context);
+            builder.defaultMiningSpeed(speed);
         }
 
         if (this.damagePerBlock != null) {
